+++ packages/transaction-controller/src/utils/batch.test.ts
@@ -18,6 +18,7 @@ import {
 } from './feature-flags';
 import { simulateGasBatch } from './gas';
 import { validateBatchRequest } from './validation';
+import type { TransactionControllerState } from '..';
 import {
   TransactionEnvelopeType,
   type TransactionControllerMessenger,
@@ -1441,7 +1442,7 @@ describe('Batch Utils', () => {
           ...request,
           publishBatchHook: undefined,
           messenger: MESSENGER_MOCK,
-          request: { ...request.request, useHook: true, origin },
+          request: { ...request.request, useHook: true, origin: ORIGIN_MOCK },
         }).catch(() => {
           // Intentionally empty
         });
@@ -1454,7 +1455,7 @@ describe('Batch Utils', () => {
           'ApprovalController:addRequest',
           expect.objectContaining({
             id: expect.any(String),
-            origin: 'http://localhost',
+            origin: ORIGIN_MOCK,
             requestData: { txBatchId: expect.any(String) },
             expectsResult: true,
             type: 'transaction_batch',
@@ -1466,6 +1467,83 @@ describe('Batch Utils', () => {
 
         const result = await resultPromise;
         expect(result?.batchId).toMatch(/^0x[0-9a-f]{32}$/u);
+      });
+
+      it('saves a transaction batch and then cleans the state', async () => {
+        const { approve } = mockRequestApproval(MESSENGER_MOCK, {
+          state: 'approved',
+        });
+        mockSequentialPublishBatchHookResults();
+        setupSequentialPublishBatchHookMock(() => sequentialPublishBatchHook);
+
+        const resultPromise = addTransactionBatch({
+          ...request,
+          publishBatchHook: undefined,
+          messenger: MESSENGER_MOCK,
+          request: {
+            ...request.request,
+            useHook: true,
+            origin: ORIGIN_MOCK,
+          },
+        }).catch(() => {
+          // Intentionally empty
+        });
+
+        await flushPromises();
+        approve();
+        await executePublishHooks();
+
+        expect(MESSENGER_MOCK.call).toHaveBeenCalledWith(
+          'ApprovalController:addRequest',
+          expect.objectContaining({
+            id: expect.any(String),
+            origin: ORIGIN_MOCK,
+            requestData: { txBatchId: expect.any(String) },
+            expectsResult: true,
+            type: 'transaction_batch',
+          }),
+          true,
+        );
+
+        expect(updateMock).toHaveBeenCalledTimes(2);
+        expect(updateMock).toHaveBeenCalledWith(expect.any(Function));
+
+        // Simulate the state update
+        const state = {
+          transactionBatches: [],
+        } as unknown as TransactionControllerState;
+        updateMock.mock.calls[0][0](state);
+
+        expect(state.transactionBatches).toStrictEqual([
+          expect.objectContaining({
+            id: expect.any(String),
+            chainId: CHAIN_ID_MOCK,
+            networkClientId: NETWORK_CLIENT_ID_MOCK,
+            transactions: [
+              expect.objectContaining({
+                params: {
+                  data: DATA_MOCK,
+                  to: TO_MOCK,
+                  value: VALUE_MOCK,
+                },
+              }),
+              expect.objectContaining({
+                params: {
+                  data: DATA_MOCK,
+                  to: TO_MOCK,
+                  value: VALUE_MOCK,
+                },
+              }),
+            ],
+            origin: ORIGIN_MOCK,
+          }),
+        ]);
+
+        await resultPromise;
+
+        updateMock.mock.calls[1][0](state);
+
+        expect(state.transactionBatches).toStrictEqual([]);
       });
     });
   });